% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble.R
\name{assemble}
\alias{assemble}
\title{Assemble data used for survey fusion}
\usage{
assemble(
  x,
  fusion.variables = NULL,
  spatial.datasets = "all",
  window = 2,
  pca = NULL,
  replicates = FALSE
)
}
\arguments{
\item{x}{List object produced by \code{\link{prepare}}.}

\item{fusion.variables}{Character. Names of donor variables to be included in output as fusion candidates. If NULL (default), an attempt is made to return all donor variables not used in harmonization process.}

\item{spatial.datasets}{Character. Vector of requested spatial datasets to merge (e.g. \code{"EPA-SLD"}) or either of two special values: \code{"all"} (default) or \code{"none"}.}

\item{window}{Integer. Size of allowable temporal window, in years, when merging spatial variables. \code{window = 0} (default) means that a spatial variable is only included if it has the same vintage as the survey. See Details.}

\item{pca}{Numeric. Controls whether/how PCA is used to reduce dimensionality of spatial variables. Default (NULL) is no PCA. If non-NULL, should be a numeric vector of length two; e.g. \code{pca = c(50, 0.95)}. First number is the maximum number of components to return; second number is target proportion of variance explained. See Details.}

\item{replicates}{Logical. Should replicate observation weights be included, if available? Defaults to FALSE. Passed to \link{completeDonor}.}
}
\value{
A list of length two containing donor and recipient microdata to pass to \code{\link[fusionModel]{train}} and \code{\link[fusionModel]{fuse}}.
}
\description{
Assembles data inputs to pass to \code{\link[fusionModel]{train}} and \code{\link[fusionModel]{fuse}} to perform survey fusion. Adds fusion, replicate weight, and/or spatial variables and checks that donor and recipient output data frames are consistent.
}
\details{
Spatial variables are included if the associated vintage is within +/- \code{window} years of the survey vintage. In cases where the spatial variable has multiple vintages equidistant from the survey vintage, the older vintage is selected. Variables with \code{vintage = "always"} are, of course, always included.

PCA is restricted to numeric spatial variables and is computed using \code{\link[stats]{prcomp}}. The returned number of principal components is the lesser of \code{pca[1]} or the number of components that explain at least \code{pca[2]} proportion of the variance. For example, \code{pca = c(50, 0.95)} will select the fewest number of components that explain 95\% of the variance, up to 50 components maximum. NA's in numeric spatial variables are imputed using median value prior to computing the principal components.
}
\examples{
prep <- prepare(donor = "RECS_2015",
                recipient = "ACS_2015",
                respondent = "household",
                implicates = 3)

data <- assemble(x = prep)

}
