% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignLocation.R
\name{assignLocation}
\alias{assignLocation}
\title{Impute or assign location for survey records}
\usage{
assignLocation(harmonized, m = 1)
}
\arguments{
\item{harmonized}{Output from call to \link{harmonize}.}

\item{m}{Integer. Number of implicates when imputing PUMA for donor observations.}
}
\value{
A list of length 2.
}
\description{
Imputes PUMA for donor microdata observations. Assigns available location variables to recipient microdata observations. Used with \link{prepare}. ADD MORE.
}
