% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BEA_pce_national}
\alias{BEA_pce_national}
\title{National Personal Consumption Expenditures (PCE)}
\format{
A \code{data.table} data frame with 73 variables:
\describe{
\item{line}{BEA line item number (integer)}
\item{pce_desc}{Description of the PCE line item category (character)}
\item{pce_series}{BEA series identification code (character)}
\item{parent1}{Hierarchical parent series ID level 1 (character)}
\item{parent2}{Hierarchical parent series ID level 2 (character)}
\item{parent3}{Hierarchical parent series ID level 3 (character)}
\item{parent4}{Hierarchical parent series ID level 4 (character)}
\item{parent5}{Hierarchical parent series ID level 5 (character)}
\item{parent6}{Hierarchical parent series ID level 6 (character)}
\item{parent7}{Hierarchical parent series ID level 7 (character)}
\item{parent8}{Hierarchical parent series ID level 8 (character)}
\item{1959-2020}{Annual PCE expenditure/index values by year column (integer)}
}
}
\source{
U.S. Bureau of Economic Analysis (BEA).
}
\usage{
BEA_pce_national
}
\description{
Bureau of Economic Analysis (BEA) national-level Personal Consumption Expenditures price index and series from 1959 to 2020 used for macroeconomic deflating and scaling.
}
\keyword{internal}
