% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BEA_pce_state}
\alias{BEA_pce_state}
\title{State Personal Consumption Expenditures (PCE)}
\format{
A \code{data.table} data frame with 31 variables:
\describe{
\item{line}{BEA line item number (integer)}
\item{state_fips}{2-digit state FIPS code (character)}
\item{state_name}{State name (character)}
\item{pce_desc}{Description of the PCE line item category (character)}
\item{pce_series}{BEA series identification code (character)}
\item{parent1}{Hierarchical parent series ID level 1 (character)}
\item{parent2}{Hierarchical parent series ID level 2 (character)}
\item{parent3}{Hierarchical parent series ID level 3 (character)}
\item{1997-2019}{Annual state-level PCE expenditure/index values by year column (numeric/integer)}
}
}
\source{
U.S. Bureau of Economic Analysis (BEA).
}
\usage{
BEA_pce_state
}
\description{
Bureau of Economic Analysis (BEA) state-level Personal Consumption Expenditures price index and series from 1997 to 2019.
}
\keyword{internal}
