% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bg_centroids}
\alias{bg_centroids}
\title{Block group centroids circa 2010}
\format{
A \code{sf} spatial data frame.
\describe{
\item{state}{State code}
\item{county10}{County code}
\item{tract10}{Tract code}
\item{bg10}{Block group code}
\item{geometry}{Centroid coordinates as \code{`sfc_POINT`} class}
}
}
\source{
\url{https://www2.census.gov/geo/docs/reference/cenpop2010/blkgrp/}
}
\usage{
bg_centroids
}
\description{
Population-weighted block group centroids. \code{sf} points object. Useful for assigning any coordinate-based spatial features to block groups to create geographic concordance; e.g. using \code{\link[sf]{st_nearest_feature}}.
}
\keyword{datasets}
