% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bg_centroids_2010}
\alias{bg_centroids_2010}
\title{Block group centroids circa 2010}
\format{
A \code{sf} spatial data frame with 6 variables:
\describe{
\item{state}{State FIPS code (character)}
\item{county10}{2010 County FIPS code (character)}
\item{tract10}{2010 Census tract code (character)}
\item{bg10}{2010 Block group code (character)}
\item{pop10}{2010 total population count (integer)}
\item{geometry}{Centroid coordinates as \code{sfc_POINT} class}
}
}
\source{
\url{https://www2.census.gov/geo/docs/reference/cenpop2010/blkgrp/}
}
\usage{
bg_centroids_2010
}
\description{
Population-weighted block group centroids. \code{sf} points object. Useful for assigning any coordinate-based spatial features to block groups to create geographic concordance; e.g. using \code{\link[sf]{st_nearest_feature}}.
}
\keyword{internal}
