% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeDonor.R
\name{completeDonor}
\alias{completeDonor}
\title{Add variables to harmonized donor microdata}
\usage{
completeDonor(data, ..., replicates = FALSE)
}
\arguments{
\item{data}{Data frame. Donor microdata, usually produced by \link{harmonize}.}

\item{...}{Optional arguments passed to \code{\link[dplyr]{select}} for selecting (or excluding) particular donor variables.}

\item{replicates}{Logical. Should replicate observation weights be included, if available? Defaults to FALSE.}
}
\value{
A data frame containing all of the harmonized variables, along with all or some of the original (unharmonized) donor variables.
}
\description{
Add donor variables to (i.e. complete) harmonized microdata, excluding variables that are harmonized with the recipient. This is useful for augmenting harmonized donor microdata produced by \link{harmonize} in preparation for passing to \code{\link[fusionModel]{train}}.
}
\examples{

}
