% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusionData-package.R
\docType{package}
\name{fusionData-package}
\alias{fusionData}
\alias{fusionData-package}
\title{fusionData: Data backend for fusionACS platform}
\description{
Data backend for fusionACS platform
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ummel.github.io/fusionData/}
}

}
\author{
\strong{Maintainer}: Kevin Ummel \email{kevin.ummel@yale.edu}

Authors:
\itemize{
  \item Kevin Ummel \email{kevin.ummel@yale.edu}
}

Other contributors:
\itemize{
  \item Karthik Akkiraju [contributor]
  \item Hero Ashman [contributor]
  \item Cora Kingdon [contributor]
  \item Miguel Poblete Cazenave [contributor]
  \item Steven Herrera Tenorio [contributor]
}

}
\keyword{internal}
