% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSurveyRaw.R
\name{getSurveyRaw}
\alias{getSurveyRaw}
\title{Download raw microdata}
\usage{
getSurveyRaw(survey)
}
\arguments{
\item{survey}{Character. Survey identifier (e.g. "RECS")}
}
\value{
\code{\link[googledrive]{drive_download}} prints messages to console indicating which files were downloaded.
}
\description{
Since raw survey microdata files are generally too large to upload to Github or attach to the fusionData package, all of the raw data files are stored on Google Drive. This function downloads requested files and places them in the appropriate local sub-directory of /fusionData. The full file path (e.g. intermediate directories) are created, if necessary, but no existing data on disk is altered or deleted.
}
\details{
\code{survey} must include a unique survey identifier. If just the identifier is provided (e.g. \code{survey = "RECS"}), then all available processed microdata across vintages and respondents is downloaded. Alternatively, can specify survey and vintage (e.g. \code{survey = "RECS_2015"}) or include additional respondent code (e.g. \code{survey = "RECS_2015_H"}) to limit download to more restrictive subsets. The special argument \code{survey = "all"} will download ALL available processed microdata across surveys.\cr\cr
The function will fail with an explanation if the current working directory is not set to \code{/fusionModel}; i.e. the project directory as pulled from Github.
}
\examples{
getSurveyRaw(survey = "RECS_2015_H")

}
