% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputePUMA.R
\name{imputePUMA}
\alias{imputePUMA}
\title{Impute PUMA for survey records}
\usage{
imputePUMA(survey, m = 1, safely = TRUE)
}
\value{
A data frame.
}
\description{
Imputes the PUMA associated with survey records by randomly drawing a PUMA from the geographic area defined by available geographic identifiers for a given record. \code{m} PUMA's are randomly drawn for each record, where the probability of selection is proportional to the number of housing units in each PUMA.
}
\examples{
imputePUMA(suvey = "RECS_2015", m = 1)

}
