% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{overlap}
\alias{overlap}
\title{Overlap coefficient between two Normal distributions}
\usage{
overlap(mu1, mu2, sd1, sd2, plot = FALSE)
}
\arguments{
\item{mu1}{Mean of the first distribution.}

\item{mu2}{Mean of the second distribution.}

\item{sd1}{Standard deviation of the first distribution.}

\item{sd2}{Standard deviation of the second distribution.}

\item{plot}{Logical. If \code{TRUE}, PDF's of the distributions are plotted with the overlap area highlighted.}
}
\value{
Numeric value \verb{[0, 1]} indicating extent of overlap.
}
\description{
Based on example and code \href{https://stats.stackexchange.com/questions/12209/percentage-of-overlapping-regions-of-two-normal-distributions}{provided here}.
}
\details{
Calculates the overlap coefficient, an intuitive measure of similarity between two Normal distributions.
}
\examples{
overlap(mu1 = 10, mu2 = 14, sd1 = 2, sd2 = 3, plot = TRUE)

}
\references{
\href{https://www.tandfonline.com/doi/abs/10.1080/03610928908830127}{Inman and Bradley (1989)}
}
