% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poverty_thresholds}
\alias{poverty_thresholds}
\title{Historical Official Poverty Thresholds}
\format{
A \code{data.table} data frame with 5 variables:
\describe{
\item{year}{4-digit calendar year (integer)}
\item{size}{Total family unit size (integer)}
\item{minors}{Number of related minor children under 18 years (integer)}
\item{senior}{Logical flag indicating if the householder is 65 years or older for single and two-person units (logical)}
\item{threshold}{Official dollar poverty threshold (integer)}
}
}
\source{
U.S. Census Bureau, Current Population Survey (CPS) Poverty Thresholds.
}
\usage{
poverty_thresholds
}
\description{
Historical official U.S. Census Bureau poverty thresholds organized by family size, number of related minor children, senior status, and year.
}
\keyword{internal}
