% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{puma_crosswalk}
\alias{puma_crosswalk}
\title{PUMA Geographic Crosswalk}
\format{
A data frame with 4 variables:
\describe{
\item{state}{2-digit state FIPS code (character)}
\item{puma20}{2020 5-digit PUMA code (character)}
\item{puma10}{2010 5-digit PUMA code (character)}
\item{xwalk_weight}{Population (2020) of the geographic intersection (integer)}
}
}
\source{
2010-2020 PUMA NHGIS crosswalk (https://usa.ipums.org/usa/volii/pumas20.shtml)
}
\usage{
puma_crosswalk
}
\description{
Geographic relationship and crosswalk table linking Public Use Microdata Areas (PUMAs) across 2010 and 2020 Census definitions.
}
\keyword{internal}
