% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadFiles.R
\name{uploadFiles}
\alias{uploadFiles}
\title{Upload local fusionData files to remote Google Drive storage}
\usage{
uploadFiles(files, ask = TRUE)
}
\arguments{
\item{files}{Character. Local file path(s) to be uploaded. Must be within local /fusionData directory.}

\item{ask}{Logical. Should user be prompted to confirm upload decision in console? Automatically set to \code{FALSE} if \code{interactive() == FALSE}.}
}
\value{
Local \code{files} uploaded to Google Drive account. Messages printed to console confirming actions.
}
\description{
Convenient and safe wrapper around \code{\link[googledrive]{googledrive}} functions to upload local file(s) in /fusionData to analogous location in /fusionData directory of "fusionACSdata" Google Drive account.
}
