% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bg_crosswalk}
\alias{bg_crosswalk}
\title{Block Group Geographic Crosswalk}
\format{
A data frame with 3 variables:
\describe{
\item{bg10}{2010 12-digit block group GEOID (character)}
\item{bg20}{2020 12-digit block group GEOID (character)}
\item{xwalk_weight}{Population (2020) of the geographic intersection (integer)}
}
}
\source{
2020 Block Groups to 2010 Block Groups NHGIS crosswalk (https://www.nhgis.org/geographic-crosswalks)
}
\usage{
bg_crosswalk
}
\description{
Geographic relationship and crosswalk table linking block groups across 2010 and 2020 Census definitions.
}
\keyword{internal}
