% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileDictionary.R
\name{compileDictionary}
\alias{compileDictionary}
\title{Compile universal survey dictionary}
\usage{
compileDictionary()
}
\value{
Saves \code{dictionary.rds} and \code{surveys.rds} data frames to disk.
}
\description{
Compiles all of the individual survey data dictionaries in \code{/survey-processed} into a single tibble that is saved to disk for use by \link{dictionary} and \link{harmony} Shiny apps.
}
\examples{
compileDictionary()

}
