% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileDictionary.R
\name{compileDictionary}
\alias{compileDictionary}
\title{Compile universal survey dictionary}
\usage{
compileDictionary()
}
\value{
Saves \code{dictionary.rda} and \code{surveys.rda} data frames to disk in \code{/data}, \code{/universe/www}, and \code{/harmony/www}.
}
\description{
Compiles all of the individual survey data dictionaries in \code{/survey-processed} into two tibbles that are saved to disk for use by fusionData package as well as \link{universe} and \link{harmony} Shiny apps.
}
\examples{
compileDictionary()

}
