% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileSpatial.R
\name{compileSpatial}
\alias{compileSpatial}
\title{Compile all spatial predictor variables}
\usage{
compileSpatial()
}
\value{
Saves \code{/geo-processed/geo_predictors.fst} to disk.
}
\description{
Detects and compiles all processed spatial datasets located in \code{/geo-processed} into a single .fst file to be utilized by \link{prepare}.
}
\examples{
compileSpatial()

}
