% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conveyHarmony.R
\name{conveyHarmony}
\alias{conveyHarmony}
\title{Convey an existing harmony file to new survey(s)}
\usage{
conveyHarmony(from, to, overwrite = FALSE)
}
\arguments{
\item{from}{Character. Name of the .R harmony file with existing harmonies of interest (e.g. "RECS_2015__ACS_2019.R")}

\item{to}{Character. Name of desired new harmony file (e.g. "RECS_2015__ACS_2015.R").}

\item{overwrite}{Logical. If \code{to} already exists, should it be overwritten? Default is \code{FALSE}.}
}
\value{
The \code{to} harmony file is written to disk (at \code{/harmony/harmonies}), if valid harmonies are detected. The list object with retained harmonies is written to disk using internal function at \code{harmony/R/harmony2dotR.R}. Possible message printed to console indicating harmonies that are not strictly valid but should be explored manually via \link{harmony} app.
}
\description{
Attempts to convey harmonies in an existing harmony file to a new harmony file for different survey(s). This is useful when introducing a new vintage of an existing survey, and some of the already-specified harmonies may be valid for the new vintage.
}
\details{
For a harmony in \code{from} to be considered valid for \code{to}, it must match exactly on variable names and all factor levels (if present). User should always manually check the \code{to} .R file before using it for analysis!
}
\examples{
conveyHarmony(from = "RECS_2015__ACS_2019.R", to = "RECS_2015__ACS_2015.R")

}
