% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datasets}
\alias{datasets}
\title{Internal datasets available in fusionData}
\description{
An overview of internal reference datasets bundled with the package.
}
\details{
\describe{
\item{\code{\link{bg_centroids_2010}}}{Population-weighted block group centroids (2010 Census).}
\item{\code{\link{bg_centroids_2020}}}{Population-weighted block group centroids (2020 Census).}
\item{\code{\link{bg_crosswalk}}}{Geographic crosswalk between 2010 and 2020 block groups.}
\item{\code{\link{BEA_pce_national}}}{BEA national Personal Consumption Expenditures series (1959–2020).}
\item{\code{\link{BEA_pce_state}}}{BEA state-level Personal Consumption Expenditures series (1997–2019).}
\item{\code{\link{cpi_series}}}{Annual Consumer Price Index series from FRED.}
\item{\code{\link{poverty_thresholds}}}{Historical official U.S. Census Bureau poverty thresholds.}
\item{\code{\link{puma_crosswalk}}}{Geographic crosswalk between 2010 and 2020 PUMAs.}
}
}
\keyword{datasets}
