% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{dictionary}
\alias{dictionary}
\title{Open the fusionACS Universal Survey Dictionary}
\usage{
dictionary()
}
\value{
Opens the app in a new browser window. R session is occupied while that window is open.
}
\description{
Opens the 'dictionary' Shiny app (/dictionary); i.e. fusionACS Universal Survey Dictionary.
}
\examples{
dictionary()

}
