% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusionData-package.R
\docType{package}
\name{fusionData-package}
\alias{fusionData}
\alias{fusionData-package}
\title{fusionData: Data backend for fusionACS platform}
\description{
Pre-process raw survey and spatial data. Harmonize variables across surveys. Build harmonized microdata for use in fusionModel package.
}
\author{
\strong{Maintainer}: Kevin Ummel \email{ummel@sas.upenn.edu}

}
\keyword{internal}
