% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusionInput.R
\name{fusionInput}
\alias{fusionInput}
\title{Generate /input files needed for fusion}
\usage{
fusionInput(
  donor,
  recipient,
  respondent,
  fuse = NULL,
  force = NULL,
  notes = NULL,
  test_mode = TRUE,
  ncores = getOption("fusionData.cores")
)
}
\arguments{
\item{donor}{Character. Donor survey identifier (e.g. \code{"RECS_2015"}).}

\item{recipient}{Character. Recipient (ACS) survey identifier (e.g. \code{"ACS_2015"}).}

\item{respondent}{Character. Desired respondent level of microdata. Either \code{"household"} or \code{"person"}.}

\item{fuse}{Character or list. Names of donor variables to be fused to recipient. If \code{fuse} is a list, each entry is a character vector possibly indicating multiple variables to fuse as a block. The order of the \code{fuse} variables does not matter, since \code{\link[fusionModel]{prepXY}} is used internally to determine a plausible fusion sequence. If NULL (default), an attempt is made to return all donor variables not used in predictor harmonization process.}

\item{force}{Character. Pre-specified subset of potential predictor variables to "force" as included predictors. These variables are also used within \code{\link{fusionOutput}} to create validation subsets. We generally select the variables that best reflect the following socioeconomic and geographic concepts: income; race/ethnicity; education; household size; housing tenure; and the highest-resolution location variable for which the donor survey is thought to be representative.}

\item{notes}{Character. Optional notes supplied by user. These are added to the log file for reference.}

\item{test_mode}{Logical. If \code{TRUE} (default), function uses the local "/fusion_" sub-directory (creating it if necessary). Only when \code{test_mode = FALSE} is it possible to overwrite production data in /fusion (no underscore).}

\item{ncores}{Integer. Number of physical CPU cores used for parallel computation.}
}
\value{
Saves resulting /input data files to appropriate local directory and (optionally) to remote Google Drive storage. Also saves a .txt log file alongside data files that records console output from \code{fusionInput()}.
}
\description{
Handles all operations needed to generate /input files from successfully ingested and harmonized donor survey microdata. Also (optionally) uploads resulting local /input data files to correct location in remote storage (via \code{\link{uploadFiles}}).
\cr\cr NOTE: Argument \code{test_mode} is \code{TRUE} by default. It must be explicitly set to \code{FALSE} to create (or possibly overwrite) production data in local /fusion directory.
}
\details{
The following steps:
}
