% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusionOutput.R
\name{fusionOutput}
\alias{fusionOutput}
\title{Generate output files resulting from fusion}
\usage{
fusionOutput(
  input,
  output = NULL,
  M = NULL,
  note = NULL,
  test_mode = TRUE,
  upload = FALSE,
  ncores = getOption("fusionData.cores"),
  margin = 2,
  ...
)
}
\arguments{
\item{input}{Character. Path to directory containing files created by \code{fusionInput}.}

\item{output}{Character. Optional path to directory where output files will be saved. If \code{output = NULL} (default), the output directory is automatically constructed from \code{input}.}

\item{M}{Integer. Desired number of fusion implicates. If \code{M = NULL} (default) it is internally set to 40 or, if \code{test_mode = TRUE}, 2 implicates.}

\item{note}{Character. Optional note supplied by user. Inserted in the log file for reference.}

\item{test_mode}{Logical. If \code{test_mode = TRUE} (default), the result files are always saved within "/fusion_", faster hyperparameters are used for \code{\link[fusionModel]{train}}, and the internal validation step is skipped.}

\item{upload}{Logical. Should resulting local /output data files be uploaded to analogous location in remote Google Drive storage? Default is \code{FALSE}.}

\item{ncores}{Integer. Number of physical CPU cores used for parallel computation.}

\item{margin}{Numeric. Passed to same argument in \code{\link[fusionModel]{fuse}}.}

\item{...}{Optional, non-default arguments passed to \code{\link[fusionModel]{train}}. For example, \code{fork = TRUE} to enable forked parallel processing.}
}
\value{
Saves resulting /output data files to appropriate local directory and (optionally) to remote Google Drive storage. Also saves a .txt log file alongside data files that records console output from \code{fusionOutput}.
}
\description{
Handles all operations needed to "do fusion" using input files generated by a successful call to \code{fusionInput}. Trains a fusion model, generates internal validation results, and then simulates multiple implicates for recipient microdata. Also (optionally) uploads resulting local /output data files to correct location in remote storage (via \code{\link{uploadFiles}}).
}
\details{
The function checks arguments and determines the file path to the appropriate /output directory (creating it if necessary). The output files are always placed within the appropriate directory hierarchy, based on the donor and recipient information detected in the \code{input} file names. In practice, \code{output} need only be specified if working in an environment where the output files need to located somewhere different from the input files.

The function executes the following steps:
\enumerate{
\item \strong{Load training data inputs}. Loads donor training microdata and results of \code{\link[fusionModel]{prepXY}}.
\item \strong{Run fusionModel::train()}. Calls \code{\link[fusionModel]{train}} using sensible defaults and hyperparameters. If \code{test_mode = TRUE}, the hyperparameters are designed to do a fast/rough-and-ready model training.
\item \strong{Fuse onto training data for internal validation}. Fuses multiple implicates to original donor training data using \code{\link[fusionModel]{fuse}}. Results saved to disk. If \code{test_mode = TRUE}, this step is skipped.
\item \strong{Run fusionModel::validate()}. Passes previous step's results to \code{\link[fusionModel]{validate}}. Results saved to disk. If \code{test_mode = TRUE}, this step is skipped.
\item \strong{Fuse onto prediction data}. Fuses multiple implicates to supplied input prediction data using \code{\link[fusionModel]{fuse}}. Results saved to disk.
\item \strong{Upload /output files to Google Drive}. If \code{upload = TRUE} and other requirements are met, uploads output data files to the analogous location in the remote Google Drive storage.
\item \strong{fusionOutput() is finished!} Upon completion, a log file named \code{"outputlog.txt"} is written to \code{/output} for reference.
}
}
\examples{
# Since 'test_mode = TRUE' by default, this will affect files in local '/fusion_' directory
dir <- fusionInput(donor = "RECS_2015",
                   recipient = "ACS_2015",
                   respondent = "household",
                   fuse = c("btung", "btuel", "cooltype"),
                   force = c("moneypy", "householder_race", "education", "nhsldmem", "kownrent", "recs_division"),
                   note = "Hello world. Reminder: running in test mode by default.")
list.files(dir)

# Using default settings: test mode and no upload to Google Drive
out <- fusionOutput(input = dir)
list.files(out)

}
