% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeoProcessed.R
\name{getGeoProcessed}
\alias{getGeoProcessed}
\title{Download processed spatial data}
\usage{
getGeoProcessed(dataset = "essential")
}
\arguments{
\item{dataset}{Character. Indicate which spatial dataset(s) to download. If \code{dataset = "essential"}, only files necessary to call \link{prepare} and \link{assemble} are downloaded. If \code{dataset = "all"}, all processed spatial data files are downloaded. Otherwise, a single spatial dataset identifier (e.g. "EPA-SLD") to download all associated "*_processed.rds" files.}
}
\value{
\code{\link[googledrive]{drive_download}} prints messages to console indicating which files were downloaded.
}
\description{

}
\details{
Files are automatically placed in appropriate sub-directory of fusionData/geo_processed. When \code{dataset = "essential"}, the downloaded files are: "geo_predictors.fst" and "concordance/geo_concordance.fst".
}
\examples{
getGeoProcessed(dataset = "essential")

}
