% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeoProcessed.R
\name{getGeoProcessed}
\alias{getGeoProcessed}
\title{Download processed spatial data}
\usage{
getGeoProcessed(dataset = "essential")
}
\arguments{
\item{dataset}{Character. Indicate which spatial dataset(s) to download. If \code{dataset = "essential"}, only files necessary to call \link{prepare} are downloaded. If \code{dataset = "all"}, all processed spatial data files are downloaded; i.e. sufficient to call \link{prepare} and/or \link{compileSpatial}. Otherwise, a single spatial dataset identifier (e.g. "EPA-SLD") to download all associated "*_processed.rds" files.}
}
\value{
\code{\link[googledrive]{drive_download}} prints messages to console indicating which files were downloaded.
}
\description{

}
\details{
The function will fail with an explanation if the current working directory is not set to \code{/fusionData}; i.e. the project directory as pulled from Github. Files are automatically placed in appropriate sub-directory of fusionData/geo_processed. When \code{dataset = "essential"}, the downloaded files are: "geo_predictors.fst", "concordance/geo_concordance.fst", and "concordance/bg_centroids.rds".
}
\examples{
getGeoProcessed(dataset = "essential")

}
