% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize.R
\name{harmonize}
\alias{harmonize}
\title{Create harmonized microdata}
\usage{
harmonize(harmony.file, respondent, output = "both")
}
\arguments{
\item{harmony.file}{Character. Name of a .R harmony file located at \code{/harmony/harmonies}.}

\item{respondent}{Character. Should the output microdata be at the \code{"household"} or \code{"person"} level?}

\item{output}{Character. Can be \code{"both"}, \code{"donor"}, or \code{"recipient"}, indicating which microdata to return.}
}
\value{
When \code{output = "both"}, a list of length 2 containing the donor and recipient data frames. Otherwise, a single data frame.
}
\description{
Parses a specified .R "harmony file" produced by the Survey Harmonization Tool \link{harmony} to produce associated microdata with all available harmonized variables. This function is called within \link{prepare} and likely never needs to be called for the user directly.
}
\examples{
test <- harmonize(harmony.file = "CEI_2015-2019__ACS_2015.R", respondent = "household")

}
