% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmony.R
\name{harmony}
\alias{harmony}
\title{Open the fusionACS Survey Harmonization Tool}
\usage{
harmony()
}
\value{
Opens the app in a new browser window. R session is occupied while that window is open.
}
\description{
Opens the 'harmony' Shiny app (/harmony); i.e. fusionACS Survey Harmonization Tool.
}
\examples{
harmony()

}
