% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installPackage.R
\name{installPackage}
\alias{installPackage}
\title{Install fusionData package locally}
\usage{
installPackage()
}
\description{
Safe and convenient way to (re-)install the fusionData package locally. Compiles local survey dictionary files (\code{\link{compileDictionary}}); updates package documentation and namespace (\code{\link[roxygen2]{roxygenise}}); and then (re-)installs the package locally (\code{\link[devtools]{install}}). Missing package dependencies are installed by default; user is asked in console if they wish to upgrade already-installed dependencies.
}
\examples{
installPackage()

}
