% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare}
\alias{prepare}
\title{Prepare microdata inputs for assembly}
\usage{
prepare(donor, recipient, respondent, ncores = getOption("fusionData.cores"))
}
\arguments{
\item{donor}{Character. Donor survey identifier (e.g. \code{"RECS_2015"}).}

\item{recipient}{Character. Recipient (ACS) survey identifier (e.g. \code{"ACS_2015"}).}

\item{respondent}{Character. Desired respondent level of microdata. Either \code{"household"} or \code{"person"}.}

\item{ncores}{Integer. Number of physical CPU cores used for parallel computation.}
}
\value{
A list of length two containing output data frames specific to donor and recipient output, respectfully. Can be passed to \code{\link{assemble}}.
}
\description{
Prepares data inputs to pass to \code{\link{assemble}}. Harmonizes common variables for the specified donor and recipient surveys, imputes PUMA for donor records, and samples location variables for recipient records.
}
\examples{
prep <- prepare(donor = "RECS_2015",
                recipient = "ACS_2015",
                respondent = "household",
                implicates = 3)

}
